+++ MakerFarm/Models/MaterialModel.cs
@@ -7,7 +7,7 @@ namespace MakerFarm.Models
     {
         public long Id { get; set; } /* Generic Id for the Material */
         public string MaterialName { get; set; } /* The Material Name (IE, Navy Blue ABS) */
-        public Int16 PrinterTypeId { get; set; } /* The ID of the Printer Type this Material is compatible with  */
+        public int PrinterTypeId { get; set; } /* The ID of the Printer Type this Material is compatible with  */
         public long MaterialSpoolQuantity { get; set; } /* This denotes how many spools of the material are on hand, and thus the maximum number of simultanious jobs using said material */
     }
